// packages/core/varity-sdk/src/core/types.d.ts
export type Network = 'arbitrum-sepolia' | 'arbitrum-l3-testnet' | 'arbitrum-l3-mainnet';
export interface NetworkConfig {
    chainId: number;
    rpcUrl: string;
    contracts: ContractAddresses;
    explorerUrl: string;
}
export interface ContractAddresses {
    MerchantRegistry: string;
    TransactionVault: string;
    RepPerformance: string;
    ResidualCalculator: string;
    AccessControlRegistry: string;
    DataProofRegistry: string;
    VarityWalletFactory: string;
}